requirements.txt
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
mutagen>=1.47.0
Pillow>=10.0.0

# Web UI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
itsdangerous>=2.1.2

